feat/refactor: contract statistic by ids view
